Refactor:  create data driven filter list
