+++ src/components/common/ColonyMembers/MembersFilter/index.ts
@@ -1,7 +1,1 @@
-export {
-  default,
-  FormValues,
-  MemberType,
-  VerificationType,
-  BannedStatus,
-} from './MembersFilter';
+export { default, FormValues } from './MembersFilter';
