+++ src/components/common/ColonyMembers/MembersFilter/types.ts
@@ -1,0 +1,10 @@
+import { MessageDescriptor } from 'react-intl';
+
+export interface Appearance {
+  theme?: 'default' | 'alt' | 'grey' | 'grid';
+}
+
+export interface SelectOption {
+  label: MessageDescriptor | string;
+  value: string;
+}
